feat(wallet-sdk): own the cashu crypto + mint queries + wallet init (connection-core step 4a)
The SDK now owns the cashu connection layer the AccountRepository depends on.

- src/cashu.ts (split out of features/shared/cashu.ts, bodies verbatim): the locking
  derivation path, tokenToMoney, CashuCryptography + seed/xpub/privateKey queryOptions +
  getCashuCryptography (opensecret + @scure), getTokenHash, getMintAuthProvider, the mint
  info/keysets/keys queryOptions + keys, decodeCashuToken, and getInitializedCashuWallet
  (offline handling + keychain cache load). queryOptions() helper calls converted to plain
  query-options objects (typed identity — runtime-identical).
- src/agicash-mint-auth-provider.ts (moved verbatim): the NUT-21 CAT AuthProvider; its
  module-load getQueryClient capture + isLoggedIn now resolve inside the SDK.
- src/performance.ts (new): a minimal instrumentation seam — measureOperation delegates to a
  registered measurer (default pass-through); entry.client.tsx registers the web's
  Sentry-backed implementation at startup so the moved operations keep emitting the same
  Sentry spans (no server-side callers of the moved fns exist today).
- web keeps in features/shared/cashu.ts what is genuinely web: the env-derived
  cashuMintValidator (import.meta.env blocklist) and the useCashuCryptography React hook,
  plus a re-export of the SDK module (all consumers untouched).
- computeSHA256 -> @agicash/utils/sha256 (pure Web Crypto; web lib/sha256.ts is a shim;
  used by the moved getTokenHash and later by spark).
- SDK gains @scure/bip32 + @scure/bip39 (hoisted to the catalog; web flips to catalog:).

typecheck, biome, build all green; 138 tests; no direct react/DOM in SDK src beyond the
documented opensecret/localStorage exception.
