+++ apps/web-wallet/app/features/shared/cashu.ts
@@ -1,168 +1,19 @@
+// The framework-free cashu connection layer (crypto queryOptions, mint
+// queryOptions, wallet init) moved to @agicash/wallet-sdk; the re-export is
+// removed in the import-cleanup PR. The env-derived mint validator and the
+// React hook below stay in the web app.
 import {
-  getPrivateKey as getMnemonic,
-  getPrivateKeyBytes,
-} from '@agicash/opensecret';
-import {
-  type AuthProvider,
-  type GetKeysResponse,
-  type GetKeysetsResponse,
-  KeyChain,
-  Mint,
-  NetworkError,
-  type Token,
-  getDecodedToken,
-} from '@cashu/cashu-ts';
-import { HDKey } from '@scure/bip32';
-import { mnemonicToSeedSync } from '@scure/bip39';
-import {
-  type QueryClient,
-  queryOptions,
-  useQueryClient,
-} from '@tanstack/react-query';
+  type CashuCryptography,
+  getCashuCryptography,
+} from '@agicash/wallet-sdk/cashu';
+import { useQueryClient } from '@tanstack/react-query';
 import { useMemo } from 'react';
-import { getQueryClient } from '~/features/shared/query-client';
-import {
-  type ExtendedCashuWallet,
-  ExtendedMintInfo,
-  type MintPurpose,
-  extractCashuToken,
-  getCashuProtocolUnit,
-  getCashuUnit,
-  getCashuWallet,
-  sumProofs,
-} from '~/lib/cashu';
 import {
   MintBlocklistSchema,
   buildMintValidator,
 } from '~/lib/cashu/mint-validation';
-import { encodeToken } from '~/lib/cashu/token';
-import { type Currency, type CurrencyUnit, Money } from '~/lib/money';
-import { measureOperation } from '~/lib/performance';
-import { computeSHA256 } from '~/lib/sha256';
-import { getSeedPhraseDerivationPath } from '../accounts/account-cryptography';
-import { getAgicashMintAuthProvider } from './agicash-mint-auth-provider';
 
-// Cashu-specific derivation path with hardnened indexes to derive public keys for
-// locking mint quotes and proofs. 129372 is UTF-8 for 🥜 (see NUT-13) and the other
-// 2 indexes are the coin type (0) and account (0) which can be changed to derive
-// different keys if needed. This path is "proprietary" and not part of any standard.
-// The index values are unimportant as long as they are hardened and remain constant.
-// DO NOT CHANGE THIS VALUE WITHOUT UPDATING USER'S XPUB IN THE DATABASE. IF THIS
-// IS NOT DONE, THEN WE WILL CREATE THE WRONG DERIVATION PATH WHEN GETTING PRIVATE KEYS.
-export const BASE_CASHU_LOCKING_DERIVATION_PATH = "m/129372'/0'/0'";
-
-function getCurrencyAndUnitFromToken(token: Token): {
-  currency: Currency;
-  unit: CurrencyUnit;
-  formatUnit: 'sat' | 'usd';
-} {
-  if (token.unit === 'sat') {
-    return { currency: 'BTC', unit: 'sat', formatUnit: 'sat' };
-  }
-  if (token.unit === 'usd') {
-    return { currency: 'USD', unit: 'cent', formatUnit: 'usd' };
-  }
-  throw new Error(`Invalid token unit ${token.unit}`);
-}
-
-export function tokenToMoney(token: Token): Money {
-  const { currency, unit } = getCurrencyAndUnitFromToken(token);
-  const amount = sumProofs(token.proofs);
-  return new Money<Currency>({
-    amount,
-    currency,
-    unit,
-  });
-}
-
-export type CashuCryptography = {
-  getSeed: () => Promise<Uint8Array>;
-  getXpub: (derivationPath?: string) => Promise<string>;
-  getPrivateKey: (derivationPath?: string) => Promise<string>;
-};
-
-const seedDerivationPath = getSeedPhraseDerivationPath('cashu', 12);
-
-export const seedQueryOptions = () =>
-  queryOptions({
-    queryKey: ['cashu-seed'],
-    queryFn: async () => {
-      const response = await getMnemonic({
-        seed_phrase_derivation_path: seedDerivationPath,
-      });
-      return mnemonicToSeedSync(response.mnemonic);
-    },
-    staleTime: Number.POSITIVE_INFINITY,
-  });
-
-export const xpubQueryOptions = ({
-  queryClient,
-  derivationPath,
-}: { queryClient: QueryClient; derivationPath?: string }) =>
-  queryOptions({
-    queryKey: ['cashu-xpub', derivationPath],
-    queryFn: async () => {
-      const seed = await queryClient.fetchQuery(seedQueryOptions());
-      const hdKey = HDKey.fromMasterSeed(seed);
-
-      if (derivationPath) {
-        const childKey = hdKey.derive(derivationPath);
-        return childKey.publicExtendedKey;
-      }
-
-      return hdKey.publicExtendedKey;
-    },
-    staleTime: Number.POSITIVE_INFINITY,
-  });
-
-const privateKeyQueryOptions = ({
-  derivationPath,
-}: { derivationPath?: string } = {}) =>
-  queryOptions({
-    queryKey: ['cashu-private-key', derivationPath],
-    queryFn: async () => {
-      const response = await getPrivateKeyBytes({
-        seed_phrase_derivation_path: seedDerivationPath,
-        private_key_derivation_path: derivationPath,
-      });
-      return response.private_key;
-    },
-    staleTime: Number.POSITIVE_INFINITY,
-  });
-
-/**
- * Gets Cashu cryptography functions.
- * @returns The Cashu cryptography functions.
- */
-export function getCashuCryptography(
-  queryClient: QueryClient,
-): CashuCryptography {
-  return {
-    getSeed: () => queryClient.fetchQuery(seedQueryOptions()),
-    getXpub: (derivationPath?: string) =>
-      queryClient.fetchQuery(xpubQueryOptions({ queryClient, derivationPath })),
-    getPrivateKey: (derivationPath?: string) =>
-      queryClient.fetchQuery(privateKeyQueryOptions({ derivationPath })),
-  };
-}
-
-/**
- * Hook that provides the Cashu cryptography functions.
- * Reference of the returned data is stable and doesn't change between renders.
- * @returns The Cashu cryptography functions.
- */
-export function useCashuCryptography(): CashuCryptography {
-  const queryClient = useQueryClient();
-
-  return useMemo(() => getCashuCryptography(queryClient), [queryClient]);
-}
-
-export function getTokenHash(token: Token | string): Promise<string> {
-  if (typeof token === 'string') {
-    return computeSHA256(token);
-  }
-  return computeSHA256(encodeToken(token));
-}
+export * from '@agicash/wallet-sdk/cashu';
 
 const mintBlocklist = MintBlocklistSchema.parse(
   JSON.parse(import.meta.env.VITE_CASHU_MINT_BLOCKLIST ?? '[]'),
@@ -174,183 +25,13 @@ export const cashuMintValidator = buildMintValidator({
   blocklist: mintBlocklist,
 });
 
-export function getMintAuthProvider(purpose: MintPurpose | undefined) {
-  return purpose === 'gift-card' || purpose === 'offer'
-    ? getAgicashMintAuthProvider()
-    : undefined;
-}
-
-export const mintInfoQueryKey = (mintUrl: string) => ['mint-info', mintUrl];
-export const allMintKeysetsQueryKey = (mintUrl: string) => [
-  'all-mint-keysets',
-  mintUrl,
-];
-export const mintKeysQueryKey = (mintUrl: string, keysetId?: string) => [
-  'mint-keys',
-  mintUrl,
-  keysetId,
-];
-
 /**
- * Get the mint info.
- *
- * @param mintUrl
- * @returns The mint info.
+ * Hook that provides the Cashu cryptography functions.
+ * Reference of the returned data is stable and doesn't change between renders.
+ * @returns The Cashu cryptography functions.
  */
-export const mintInfoQueryOptions = (mintUrl: string) =>
-  queryOptions({
-    queryKey: mintInfoQueryKey(mintUrl),
-    queryFn: async () =>
-      new ExtendedMintInfo(await new Mint(mintUrl).getInfo()),
-    staleTime: 1000 * 60 * 60, // 1 hour
-  });
+export function useCashuCryptography(): CashuCryptography {
+  const queryClient = useQueryClient();
 
-/**
- * Get the mints keysets in no specific order.
- *
- * @param mintUrl
- * @returns All the mints past and current keysets.
- */
-export const allMintKeysetsQueryOptions = (mintUrl: string) =>
-  queryOptions({
-    queryKey: allMintKeysetsQueryKey(mintUrl),
-    queryFn: async () => new Mint(mintUrl).getKeySets(),
-    staleTime: 1000 * 60 * 60, // 1 hour
-  });
-
-/**
- * Extract and decode a cashu token from arbitrary content.
- * Fetches keyset IDs from the token's mint for v2 keyset resolution.
- */
-export async function decodeCashuToken(content: string): Promise<Token | null> {
-  const result = extractCashuToken(content);
-  if (!result) return null;
-
-  try {
-    const queryClient = getQueryClient();
-    const data = await queryClient.fetchQuery(
-      allMintKeysetsQueryOptions(result.metadata.mint),
-    );
-    const keysetIds = data.keysets.map((k) => k.id);
-    return getDecodedToken(result.encoded, keysetIds);
-  } catch (error) {
-    console.error('Failed to decode cashu token', error);
-    return null;
-  }
-}
-
-/**
- * Get the mints public keys.
- *
- * @param mintUrl
- * @param keysetId Optional param to get the keys for a specific keyset. If not specified, the
- *   keys from all active keysets are fetched.
- * @returns An object with an array of the fetched keysets.
- */
-export const mintKeysQueryOptions = (mintUrl: string, keysetId?: string) =>
-  queryOptions({
-    queryKey: mintKeysQueryKey(mintUrl, keysetId),
-    queryFn: async () => new Mint(mintUrl).getKeys(keysetId),
-    staleTime: 1000 * 60 * 60, // 1 hour
-  });
-
-/**
- * Initializes a Cashu wallet with offline handling.
- * If the mint is offline or times out, returns a minimal wallet with isOnline: false.
- * @param queryClient - The query client to use for fetching mint data.
- * @param mintUrl - The mint URL.
- * @param currency - The currency.
- * @param bip39seed - Optional BIP39 seed for wallet initialization.
- * @returns The wallet and online status.
- */
-export async function getInitializedCashuWallet({
-  queryClient,
-  mintUrl,
-  currency,
-  bip39seed,
-  authProvider,
-}: {
-  queryClient: QueryClient;
-  mintUrl: string;
-  currency: Currency;
-  bip39seed?: Uint8Array;
-  authProvider?: AuthProvider;
-}): Promise<{ wallet: ExtendedCashuWallet; isOnline: boolean }> {
-  return measureOperation(
-    'getInitializedCashuWallet',
-    async () => {
-      let mintInfo: ExtendedMintInfo;
-      let allMintKeysets: GetKeysetsResponse;
-      let mintActiveKeys: GetKeysResponse;
-
-      try {
-        [mintInfo, allMintKeysets, mintActiveKeys] = await Promise.race([
-          Promise.all([
-            queryClient.fetchQuery(mintInfoQueryOptions(mintUrl)),
-            queryClient.fetchQuery(allMintKeysetsQueryOptions(mintUrl)),
-            queryClient.fetchQuery(mintKeysQueryOptions(mintUrl)),
-          ]),
-          new Promise<never>((_, reject) => {
-            setTimeout(() => {
-              queryClient.cancelQueries({
-                queryKey: mintInfoQueryKey(mintUrl),
-              });
-              queryClient.cancelQueries({
-                queryKey: allMintKeysetsQueryKey(mintUrl),
-              });
-              queryClient.cancelQueries({
-                queryKey: mintKeysQueryKey(mintUrl),
-              });
-              reject(new NetworkError('Mint request timed out'));
-            }, 10_000);
-          }),
-        ]);
-      } catch (error) {
-        if (error instanceof NetworkError) {
-          const wallet = getCashuWallet(mintUrl, {
-            unit: getCashuUnit(currency),
-            bip39seed: bip39seed ?? undefined,
-            authProvider,
-          });
-          return { wallet, isOnline: false };
-        }
-        throw error;
-      }
-
-      const unitKeysets = allMintKeysets.keysets.filter(
-        (ks) => ks.unit === getCashuProtocolUnit(currency),
-      );
-      const activeKeyset = unitKeysets.find((ks) => ks.active);
-
-      if (!activeKeyset) {
-        throw new Error(`No active keyset found for ${currency} on ${mintUrl}`);
-      }
-
-      const activeKeysForUnit = mintActiveKeys.keysets.find(
-        (ks) => ks.id === activeKeyset.id,
-      );
-
-      if (!activeKeysForUnit) {
-        throw new Error(
-          `Got active keyset ${activeKeyset.id} from ${mintUrl} but could not find keys for it`,
-        );
-      }
-
-      const wallet = getCashuWallet(mintUrl, {
-        unit: getCashuUnit(currency),
-        bip39seed: bip39seed ?? undefined,
-        authProvider,
-      });
-      const keyChainCache = KeyChain.mintToCacheDTO(
-        wallet.unit,
-        mintUrl,
-        unitKeysets,
-        [activeKeysForUnit],
-      );
-      wallet.loadMintFromCache(mintInfo.cache, keyChainCache);
-
-      return { wallet, isOnline: true };
-    },
-    { mintUrl, currency },
-  );
+  return useMemo(() => getCashuCryptography(queryClient), [queryClient]);
 }
