+++ apps/web-wallet/app/lib/sha256.ts
@@ -1,10 +1,2 @@
-export async function computeSHA256(message: string): Promise<string> {
-  const encoder = new TextEncoder();
-  const data = encoder.encode(message);
-  const hashBuffer = await crypto.subtle.digest('SHA-256', data);
-  const hashArray = Array.from(new Uint8Array(hashBuffer));
-  const hashHex = hashArray
-    .map((b) => b.toString(16).padStart(2, '0'))
-    .join('');
-  return hashHex;
-}
+// Transitional re-export — moved to @agicash/utils; removed in the import-cleanup PR.
+export * from '@agicash/utils/sha256';
