+++ apps/web-wallet/package.json
@@ -44,8 +44,8 @@
     "@react-router/express": "7.13.0",
     "@react-router/fs-routes": "7.13.0",
     "@react-router/node": "7.13.0",
-    "@scure/bip32": "1.7.0",
-    "@scure/bip39": "1.6.0",
+    "@scure/bip32": "catalog:",
+    "@scure/bip39": "catalog:",
     "@sentry/profiling-node": "10.42.0",
     "@sentry/react-router": "10.42.0",
     "@supabase/supabase-js": "catalog:",
