+++ package.json
@@ -11,6 +11,8 @@
       "@noble/ciphers": "1.3.0",
       "@noble/curves": "1.9.7",
       "@noble/hashes": "1.8.0",
+      "@scure/bip32": "1.7.0",
+      "@scure/bip39": "1.6.0",
       "@stablelib/base64": "2.0.1",
       "@stablelib/chacha20poly1305": "2.0.1",
       "@supabase/supabase-js": "2.95.2",
