+++ packages/utils/src/index.ts
@@ -3,3 +3,4 @@ export * from './json';
 export * from './zod';
 export * from './collections';
 export * from './ecies';
+export * from './sha256';
